LDT: HyMAP2, Support flexible grid setting (dx~=dy)

+++ ldt/params/HYMAP/read_HYMAP_basin.F90
@@ -43,7 +43,7 @@ subroutine read_HYMAP_basin(n, array)
   integer :: ftn
   integer :: c,r
   logical :: file_exists
-  integer :: iarray(LDT_rc%lnc(n),LDT_rc%lnr(n),1)
+  real    :: iarray(LDT_rc%lnc(n),LDT_rc%lnr(n),1)
 
   inquire(file=trim(HYMAP_struc(n)%basinfile), exist=file_exists)
   if(.not.file_exists) then
